+++ server/bus/__init__.py
@@ -1,0 +1,1 @@
+from .bus import bus_bp
