+++ server/customer/__init__.py
@@ -1,0 +1,1 @@
+from .customer import customer_bp
