+++ server/driver/__init__.py
@@ -1,0 +1,1 @@
+from .driver import driver_bp
