+++ server/route/__init__.py
@@ -1,0 +1,1 @@
+from .route import route_bp
